refactor(dashboard): extract useProviderConnections/Settings/Models hooks — #3501 Phase 1f (#3684)
#3501 Phase 1f: god-component 4,948->4,062 LOC. Connection state+handlers, settings, and model metadata moved into hooks/.
